fix : adjust router deleteService

+++ src/modules/Unit/services/DeleteUnitservies.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 
+import { AppError } from '@shared/errors/AppError';
 import { inject, injectable } from 'tsyringe';
-import { IUnit } from '../domain/IUnit';
 import IUnitRepository from '../domain/repository/IUnitRepository';
 
 @injectable()
@@ -11,6 +11,10 @@ export default class DeleteUnitService {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const unitExist = await this.unitRepository.getById(id);
+    if (!unitExist) {
+      throw new AppError('Unit not found');
+    }
     await this.unitRepository.delete(id);
   }
 }
